Modification du type de la cadence de tire de double en int

+++ src/main/java/universite_paris8/iut/mcheema/codesource/modele/ennemi/Ralentisseur.java
@@ -29,7 +29,7 @@ public class Ralentisseur extends Ennemi {
             if(batiment instanceof BatimentTir) {
                 if (!batimentRalenti.contains(batiment)) {
                     if (Math.sqrt((this.getX() - batiment.getX()) * (this.getX() - batiment.getX()) + (this.getY() - batiment.getY()) * (this.getY() - batiment.getY())) <= this.portee) {
-                        ((BatimentTir) batiment).setCadenceTir(((BatimentTir) batiment).getCadenceTir() * this.ralentissement);
+                        ((BatimentTir) batiment).setCadenceTir((int) (((BatimentTir) batiment).getCadenceTir() * this.ralentissement));
                         this.batimentRalenti.add((BatimentTir) batiment);
                     }
                 }
@@ -39,7 +39,7 @@ public class Ralentisseur extends Ennemi {
         for(int i = this.batimentRalenti.size() - 1;i>=0;i--) {
             BatimentTir batiment = this.batimentRalenti.get(i);
             if (Math.sqrt((this.getX() - batiment.getX()) * (this.getX() - batiment.getX()) + (this.getY() - batiment.getY()) * (this.getY() - batiment.getY()))>this.portee) {
-                batiment.setCadenceTir(batiment.getCadenceTir() / this.ralentissement);
+                batiment.setCadenceTir((int) (batiment.getCadenceTir() / this.ralentissement));
                 this.batimentRalenti.remove(i);
             }
         }
